requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
akro==0.0.6
astor==0.8.1
async-timeout==3.0.1
atari-py==0.1.15
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
cma==2.7.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
deepdiff==4.3.2
dowel==0.0.3
# enum34==1.1.10
filelock==3.0.12
future==0.18.2
gast==0.2.2
grpcio==1.28.1
gym==0.20.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.12.5
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
line-profiler==3.0.2
Markdown==3.2.1
matplotlib==3.2.1
mock==4.0.2
mpyq==0.2.5
multidict==4.7.5
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.34
opt-einsum==3.2.0
ordered-set==3.1.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
portpicker==1.3.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil
ptyprocess==0.6.0
py-spy==0.3.3
pygame==2.0.0
pyglet==1.5.0
# For macOS
# pyglet==1.5.11
Pygments==2.6.1
pyparsing==2.4.7
PyPrind==2.11.2
pyrsistent==0.16.0
PySC2==3.0.0
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
ray==0.8.4
redis==3.4.1
requests==2.23.0
s2clientprotocol==4.11.4.78285.0
s2protocol==4.11.4.78285.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
sk-video==1.1.10
soupsieve==2.0
tabulate==0.8.7
tensorboard
tensorboardX==2.0
termcolor==1.1.0
torch
torchvision
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
git+https://github.com/oxwhirl/smac.git
tmuxp
pysftp
tikzplotlib
tqdm
wandb